refactor(cddl): share smart pointer traits
